+++ tractor/ipc/_tcp.py
@@ -20,14 +20,120 @@ TCP implementation of tractor.ipc._transport.MsgTransport protocol
 from __future__ import annotations
 
 import trio
+from trio import (
+    SocketListener,
+    open_tcp_listeners,
+)
 
 from tractor.msg import MsgCodec
 from tractor.log import get_logger
-from tractor._addr import TCPAddress
 from tractor.ipc._transport import MsgpackTransport
 
 
 log = get_logger(__name__)
+
+
+class TCPAddress:
+    proto_key: str = 'tcp'
+    unwrapped_type: type = tuple[str, int]
+    def_bindspace: str = '127.0.0.1'
+
+    def __init__(
+        self,
+        host: str,
+        port: int
+    ):
+        if (
+            not isinstance(host, str)
+            or
+            not isinstance(port, int)
+        ):
+            raise TypeError(
+                f'Expected host {host!r} to be str and port {port!r} to be int'
+            )
+
+        self._host: str = host
+        self._port: int = port
+
+    @property
+    def is_valid(self) -> bool:
+        return self._port != 0
+
+    @property
+    def bindspace(self) -> str:
+        return self._host
+
+    @property
+    def domain(self) -> str:
+        return self._host
+
+    @classmethod
+    def from_addr(
+        cls,
+        addr: tuple[str, int]
+    ) -> TCPAddress:
+        match addr:
+            case (str(), int()):
+                return TCPAddress(addr[0], addr[1])
+            case _:
+                raise ValueError(
+                    f'Invalid unwrapped address for {cls}\n'
+                    f'{addr}\n'
+                )
+
+    def unwrap(self) -> tuple[str, int]:
+        return (
+            self._host,
+            self._port,
+        )
+
+    @classmethod
+    def get_random(
+        cls,
+        bindspace: str = def_bindspace,
+    ) -> TCPAddress:
+        return TCPAddress(bindspace, 0)
+
+    @classmethod
+    def get_root(cls) -> TCPAddress:
+        return TCPAddress(
+            '127.0.0.1',
+            1616,
+        )
+
+    def __repr__(self) -> str:
+        return (
+            f'{type(self).__name__}[{self.unwrap()}]'
+        )
+
+    def __eq__(self, other) -> bool:
+        if not isinstance(other, TCPAddress):
+            raise TypeError(
+                f'Can not compare {type(other)} with {type(self)}'
+            )
+
+        return (
+            self._host == other._host
+            and
+            self._port == other._port
+        )
+
+    async def open_listener(
+        self,
+        **kwargs,
+    ) -> SocketListener:
+        listeners: list[SocketListener] = await open_tcp_listeners(
+            host=self._host,
+            port=self._port,
+            **kwargs
+        )
+        assert len(listeners) == 1
+        listener = listeners[0]
+        self._host, self._port = listener.socket.getsockname()[:2]
+        return listener
+
+    async def close_listener(self):
+        ...
 
 
 # TODO: typing oddity.. not sure why we have to inherit here, but it
